разделил компонент
